app/requirements.txt
Flask==2.2.5
PyYAML==6.0
hvac==1.3.3
